# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface RequestError {
  error: string;
  code?: string;
}

export type StaffProfileRole = typeof StaffProfileRole[keyof typeof StaffProfileRole];


export const StaffProfileRole = {
  staff: 'staff',
  manager: 'manager',
  admin: 'admin',
} as const;

export type StaffProfileStatus = typeof StaffProfileStatus[keyof typeof StaffProfileStatus];


export const StaffProfileStatus = {
  pending: 'pending',
  active: 'active',
  suspended: 'suspended',
  rejected: 'rejected',
} as const;

export interface StaffProfile {
  id: string;
  email: string;
  display_name: string;
  role: StaffProfileRole;
  status: StaffProfileStatus;
  password_change_required: boolean;
  /** @nullable */
  temporary_password_issued_at: string | null;
  /** @nullable */
  first_password_set_at: string | null;
  /** @nullable */
  policy_accepted_at: string | null;
  /** @nullable */
  policy_version: string | null;
}

export interface AuthSession {
  access_token: string;
  refresh_token: string;
  expires_in: number;
  token_type: string;
}

export interface SignInRequest {
  email: string;
  password: string;
}

export type SignInResponseNext = typeof SignInResponseNext[keyof typeof SignInResponseNext];


export const SignInResponseNext = {
  'set-password': 'set-password',
  policy: 'policy',
  ready: 'ready',
} as const;

export interface SignInResponse {
  session: AuthSession;
  profile: StaffProfile;
  next: SignInResponseNext;
}

export interface SetPasswordRequest {
  /** @minLength 8 */
  password: string;
}

export type OnboardingResponseNext = typeof OnboardingResponseNext[keyof typeof OnboardingResponseNext];


export const OnboardingResponseNext = {
  'set-password': 'set-password',
  policy: 'policy',
  ready: 'ready',
} as const;

export interface OnboardingResponse {
  profile: StaffProfile;
  next: OnboardingResponseNext;
}

export type CreateStaffRequestRole = typeof CreateStaffRequestRole[keyof typeof CreateStaffRequestRole];


export const CreateStaffRequestRole = {
  staff: 'staff',
  manager: 'manager',
} as const;

export interface CreateStaffRequest {
  /**
     * @minLength 1
     * @maxLength 120
     */
  displayName: string;
  email: string;
  role: CreateStaffRequestRole;
}

export interface StaffListResponse {
  staff: StaffProfile[];
}

export interface StaffCredentialResponse {
  staff: StaffProfile;
  temporaryPassword: string;
}

export type MessageAttachmentType = typeof MessageAttachmentType[keyof typeof MessageAttachmentType];


export const MessageAttachmentType = {
  image: 'image',
  document: 'document',
} as const;

export interface MessageAttachment {
  type: MessageAttachmentType;
  /** @maxLength 240 */
  name: string;
  /** @nullable */
  mimeType?: string | null;
  /** @nullable */
  size?: number | null;
  storagePath?: string;
}

export interface CreateConversationMessageRequest {
  /** @maxLength 8000 */
  body: string;
  /**
     * @minLength 8
     * @maxLength 180
     */
  clientMessageId: string;
  attachment?: MessageAttachment | null;
}

export type PersistedStaffMessageKind = typeof PersistedStaffMessageKind[keyof typeof PersistedStaffMessageKind];


export const PersistedStaffMessageKind = {
  text: 'text',
  poll: 'poll',
  event: 'event',
  system: 'system',
} as const;

export type PersistedStaffMessageMetadata = { [key: string]: unknown };

export interface PersistedStaffMessage {
  id: string;
  conversation_id: string;
  /** @nullable */
  sender_id: string | null;
  /** @nullable */
  parent_message_id: string | null;
  kind: PersistedStaffMessageKind;
  body: string;
  metadata: PersistedStaffMessageMetadata;
  /** @nullable */
  client_message_id: string | null;
  created_at: string;
  /** @nullable */
  edited_at: string | null;
}

export interface MessageDeliveryResponse {
  message: PersistedStaffMessage;
}

export type RegisterDevicePushTokenRequestPlatform = typeof RegisterDevicePushTokenRequestPlatform[keyof typeof RegisterDevicePushTokenRequestPlatform];


export const RegisterDevicePushTokenRequestPlatform = {
  ios: 'ios',
  android: 'android',
} as const;

export interface RegisterDevicePushTokenRequest {
  expoPushToken?: string;
  platform?: RegisterDevicePushTokenRequestPlatform;
  enabled: boolean;
}

/**
 * Request was rejected
 */
export type RequestErrorResponse = RequestError;