# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: API specification
servers:
  - url: /api
    description: Base API path
tags:
  - name: health
    description: Health operations
paths:
  /healthz:
    get:
      operationId: healthCheck
      tags: [health]
      summary: Health check
      description: Returns server health status
      responses:
        "200":
          description: Healthy
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HealthStatus"
  /auth/sign-in:
    post:
      operationId: signInStaff
      tags: [staff-auth]
      summary: Sign in a staff member
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SignInRequest"
      responses:
        "200":
          description: Authenticated session and onboarding state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SignInResponse"
        "401":
          $ref: "#/components/responses/RequestError"
  /auth/session:
    get:
      operationId: getStaffSession
      tags: [staff-auth]
      summary: Get current staff onboarding state
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Current staff state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingResponse"
        "401":
          $ref: "#/components/responses/RequestError"
  /auth/password:
    post:
      operationId: setStaffPassword
      tags: [staff-auth]
      summary: Replace a temporary password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SetPasswordRequest"
      responses:
        "200":
          description: Updated onboarding state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingResponse"
        "400":
          $ref: "#/components/responses/RequestError"
  /auth/policy-acceptance:
    post:
      operationId: acceptStaffPolicy
      tags: [staff-auth]
      summary: Save policy acknowledgement
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Updated onboarding state
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/OnboardingResponse"
        "409":
          $ref: "#/components/responses/RequestError"
  /staff:
    get:
      operationId: listStaffAccess
      tags: [staff-access]
      summary: List staff accounts for an administrator
      security: [{ bearerAuth: [] }]
      responses:
        "200":
          description: Staff accounts
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffListResponse"
    post:
      operationId: createStaffAccess
      tags: [staff-access]
      summary: Create staff access and a temporary password
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateStaffRequest"
      responses:
        "201":
          description: New staff account and one-time temporary password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffCredentialResponse"
  /staff/{profileId}/reset-password:
    post:
      operationId: resetStaffTemporaryPassword
      tags: [staff-access]
      summary: Replace a staff member temporary password
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: profileId
          required: true
          schema:
            type: string
      responses:
        "200":
          description: New temporary password
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/StaffCredentialResponse"
  /conversations/{conversationId}/messages:
    post:
      operationId: createConversationMessage
      tags: [messaging]
      summary: Persist an idempotent staff message
      security: [{ bearerAuth: [] }]
      parameters:
        - in: path
          name: conversationId
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateConversationMessageRequest"
      responses:
        "201":
          description: Message was persisted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageDeliveryResponse"
        "400":
          $ref: "#/components/responses/RequestError"
        "403":
          $ref: "#/components/responses/RequestError"
  /devices/push-token:
    post:
      operationId: registerDevicePushToken
      tags: [notifications]
      summary: Register this staff device for privacy-safe message notifications
      security: [{ bearerAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterDevicePushTokenRequest"
      responses:
        "204":
          description: Device token was registered
        "400":
          $ref: "#/components/responses/RequestError"
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  responses:
    RequestError:
      description: Request was rejected
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/RequestError"
  schemas:
    HealthStatus:
      type: object
      properties:
        status:
          type: string
      required:
        - status
    RequestError:
      type: object
      properties:
        error:
          type: string
        code:
          type: string
      required: [error]
    StaffProfile:
      type: object
      properties:
        id: { type: string }
        email: { type: string }
        display_name: { type: string }
        role: { type: string, enum: [staff, manager, admin] }
        status: { type: string, enum: [pending, active, suspended, rejected] }
        password_change_required: { type: boolean }
        temporary_password_issued_at: { type: [string, "null"], format: date-time }
        first_password_set_at: { type: [string, "null"], format: date-time }
        policy_accepted_at: { type: [string, "null"], format: date-time }
        policy_version: { type: [string, "null"] }
      required: [id, email, display_name, role, status, password_change_required, temporary_password_issued_at, first_password_set_at, policy_accepted_at, policy_version]
    AuthSession:
      type: object
      properties:
        access_token: { type: string }
        refresh_token: { type: string }
        expires_in: { type: number }
        token_type: { type: string }
      required: [access_token, refresh_token, expires_in, token_type]
    SignInRequest:
      type: object
      properties:
        email: { type: string }
        password: { type: string, format: password }
      required: [email, password]
    SignInResponse:
      type: object
      properties:
        session: { $ref: "#/components/schemas/AuthSession" }
        profile: { $ref: "#/components/schemas/StaffProfile" }
        next: { type: string, enum: [set-password, policy, ready] }
      required: [session, profile, next]
    SetPasswordRequest:
      type: object
      properties:
        password: { type: string, format: password, minLength: 8 }
      required: [password]
    OnboardingResponse:
      type: object
      properties:
        profile: { $ref: "#/components/schemas/StaffProfile" }
        next: { type: string, enum: [set-password, policy, ready] }
      required: [profile, next]
    CreateStaffRequest:
      type: object
      properties:
        displayName: { type: string, minLength: 1, maxLength: 120 }
        email: { type: string }
        role: { type: string, enum: [staff, manager] }
      required: [displayName, email, role]
    StaffListResponse:
      type: object
      properties:
        staff:
          type: array
          items: { $ref: "#/components/schemas/StaffProfile" }
      required: [staff]
    StaffCredentialResponse:
      type: object
      properties:
        staff: { $ref: "#/components/schemas/StaffProfile" }
        temporaryPassword: { type: string }
      required: [staff, temporaryPassword]
    MessageAttachment:
      type: object
      properties:
        type: { type: string, enum: [image, document] }
        name: { type: string, maxLength: 240 }
        mimeType: { type: [string, "null"] }
        size: { type: [number, "null"] }
        storagePath: { type: string }
      required: [type, name]
    CreateConversationMessageRequest:
      type: object
      properties:
        body: { type: string, maxLength: 8000 }
        clientMessageId: { type: string, minLength: 8, maxLength: 180 }
        attachment:
          oneOf:
            - $ref: "#/components/schemas/MessageAttachment"
            - type: "null"
      required: [body, clientMessageId]
    PersistedStaffMessage:
      type: object
      properties:
        id: { type: string }
        conversation_id: { type: string }
        sender_id: { type: [string, "null"] }
        parent_message_id: { type: [string, "null"] }
        kind: { type: string, enum: [text, poll, event, system] }
        body: { type: string }
        metadata: { type: object, additionalProperties: true }
        client_message_id: { type: [string, "null"] }
        created_at: { type: string, format: date-time }
        edited_at: { type: [string, "null"], format: date-time }
      required: [id, conversation_id, sender_id, parent_message_id, kind, body, metadata, client_message_id, created_at, edited_at]
    MessageDeliveryResponse:
      type: object
      properties:
        message: { $ref: "#/components/schemas/PersistedStaffMessage" }
      required: [message]
    RegisterDevicePushTokenRequest:
      type: object
      properties:
        expoPushToken: { type: string }
        platform: { type: string, enum: [ios, android] }
        enabled: { type: boolean }
      required: [enabled]